backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.2
torch==2.1.1
aiohttp==3.9.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
arxiv==2.0.0
crossref-commons==0.0.9
pydantic==2.5.0
python-multipart==0.0.6
